requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
cryptography
passlib[bcrypt]
python-jose[cryptography]
